chore: migrate from nodeunit to mocha
