+++ package.json
@@ -5,14 +5,14 @@
   "description": "URL-safe base64 UUID encoder for generating 22 character slugs",
   "license": "MIT",
   "scripts": {
-    "test": "nodeunit slugid_test.js",
+    "test": "mocha",
     "test-nobuffer": "NO_BUFFER=1 yarn test"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/taskcluster/slugid.git"
   },
-  "types": "index.d.ts",
+  "types": "./src/index.d.ts",
   "bugs": {
     "url": "https://github.com/taskcluster/slugid/issues"
   },
@@ -23,6 +23,6 @@
     "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "nodeunit": "^0.11.3"
+    "mocha": "^10.2.0"
   }
 }
