+++ src/slugid.js
